fix(core): pass options to tasks runner correctly
