+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@nrwl/nx-source",
-  "version": "8.9.0",
+  "version": "8.10.0-beta.7",
   "description": "Extensible Dev Tools for Monorepos",
   "homepage": "https://nx.dev",
   "main": "index.js",
