+++ scripts/e2e-ci2.sh
@@ -19,4 +19,4 @@ jest --maxWorkers=1 ./build/e2e/run-many.test.js &&
 jest --maxWorkers=1 ./build/e2e/storybook.test.js &&
 jest --maxWorkers=1 ./build/e2e/upgrade-module.test.js &&
 jest --maxWorkers=1 ./build/e2e/web.test.js &&
-jest --maxWorkers=1 ./build/e2e/cache.test.js
+jest --maxWorkers=1 ./build/e2e/default-tasks-runner.test.js
